requirements.txt
streamlit==1.29.0
openai>=1.54.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
